requirements.txt
arrow==1.2.2
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
bleach==5.0.1
certifi==2022.5.18.1
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==2.1.1
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.50.0
django-bleach==3.0.0
django-browser-reload==1.6.0
django-summernote==0.8.20.0
django-tables2==2.4.1
django-tailwind==3.3.0
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
urllib3==1.26.9
webencodings==0.5.1
